fix(ledger): record the DRep deposit as optional and fail closed on absence
DRepRegistration.Deposit was a non-optional uint64, so a ledger state
that reports a DRep as registered without a recorded deposit yielded an
authoritative zero and UtxoValidateCertificateDeposits rejected every
deregistration supplying its real refund.

Deposit is now a *uint64, matching the contract StakeCredentialDeposit
already documents: nil is the absence of a recorded deposit, distinct
from a recorded zero. The reference ledger's DRepState always carries a
deposit, so a registration without one is an inconsistent state and
refund validation returns DRepDepositStateInconsistentError instead of
comparing the refund against zero.

In-transaction DRep registrations are keyed by the full credential, so a
key-hash and a script-hash certificate sharing a hash no longer collide.

Credential stays a Blake2b224. Widening it to a full Credential is a
breaking change for every LedgerState implementor, so DRep lookups still
cannot distinguish a same-hash key/script pair; that is deferred.
